requirements.txt
django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
opencv-python-headless==4.11.0.86
channels==4.2.2
daphne==4.1.2
Pillow==11.2.0
numpy>=1.24.0
python-dotenv==1.0.1
redis==5.0.3
mysqlclient==2.2.3

# OpenCV for RTSP streaming
opencv-python==4.7.0.72
numpy==1.24.3

# Vehicle detection dependencies
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
matplotlib>=3.2.0
seaborn>=0.11.0
PyYAML>=5.3.1
requests>=2.23.0
